Unify downloader steps into pipeline
Previously we had to call specific functions for each module and pass
lots of parameters to them. There was also a lot of repeated code to
save the state and log information at each step.
To get rid of this, we've implemented a pipeline pattern,
where each sequantial step of the downloader conforms to a set
interface. The pipeline then wraps the state management and logging
around the simple execution of each step.
This commit also splits up the DOI resolving and the PDF download.