add plugin system for doi to pdf-url mappers
Having hardcoded logic to retrieve the URL to download a PDF file for a
DOI in the PdfDownloader is not scalable. To combat this, this commit
adds a Plugin system loosely based on the Strategy pattern, which allows
the user to choose the mapping function at runtime, as well as easily
add their own mappers, which are automatically loaded into
the downloader.