skip DOI resolving if not needed
Some mappers rely on the resolved DOI to build the URL to
download the PDF, while others only need the pure DOI.
To skip the unnecessary step of resolving DOIs in these cases,
there are now 2 ABCs to base a mapper on. Depending on which Mapper
is selected, the resolving step gets skipped or run.